ugly day9

+++ intcode-computer/src/io_wrapper.rs
@@ -3,8 +3,8 @@ use std::collections::VecDeque;
 
 #[derive(Debug)]
 pub struct Io {
-    input: VecDeque<i32>,
-    output: VecDeque<i32>,
+    input: VecDeque<i64>,
+    output: VecDeque<i64>,
 }
 
 impl Io {
@@ -16,29 +16,29 @@ impl Io {
     }
 
     /// Consumes self, returning the resulting IO.
-    pub fn into_output(self) -> VecDeque<i32> {
+    pub fn into_output(self) -> VecDeque<i64> {
         self.output
     }
 
     /// Read from input
-    pub fn output_read(&mut self) -> Result<i32, Error> {
+    pub fn output_read(&mut self) -> Result<i64, Error> {
         self.output.pop_front().ok_or_else(|| Error::msg("EOF"))
     }
 
     /// Read from input
-    pub fn input_write(&mut self, value: i32) -> Result<(), Error> {
+    pub fn input_write(&mut self, value: i64) -> Result<(), Error> {
         self.input.push_back(value);
 
         Ok(())
     }
 
     /// Read from input
-    pub fn read(&mut self) -> Result<i32, Error> {
+    pub fn read(&mut self) -> Result<i64, Error> {
         self.input.pop_front().ok_or(Error::msg("EOF"))
     }
 
     /// Write to output
-    pub fn write(&mut self, value: i32) -> Result<(), Error> {
+    pub fn write(&mut self, value: i64) -> Result<(), Error> {
         self.output.push_back(value);
 
         Ok(())
